+++ frontend/views/layouts/_base.php
@@ -40,7 +40,7 @@ use yii\bootstrap\NavBar;
                 ['label' => \Yii::t('frontend', 'Signup'), 'url' => ['/user/sign-in/signup'], 'visible'=>Yii::$app->user->isGuest],
                 ['label' => \Yii::t('frontend', 'Login'), 'url' => ['/user/sign-in/login'], 'visible'=>Yii::$app->user->isGuest],
                 [
-                    'label' => Yii::$app->user->isGuest ? '' : Yii::$app->user->identity->username,
+                    'label' => Yii::$app->user->isGuest ? '' : Yii::$app->user->identity->getPublicIdentity(),
                     'visible'=>!Yii::$app->user->isGuest,
                     'items'=>[
                         [